Avances
Hace métodos para calcular:

• Entrada, Plato fuerte y postre más vendidos por muestra
• Entrada, Plato fuerte y postre con mejor calificación durante cada muestra
• Mesero con mayor calificación diaria durante cada muestra
• Ingresos en bruto del restaurante en cada una de las muestras
• Las modalidades de pago presentadas en cada una de las muestras

+++ RestaurantV3/src/models/entities/Dish.java
@@ -9,9 +9,9 @@ public class Dish {
     private String name;
     private DishType dishType;
     private double price;
-    private double cookingTime; //In minutes
+    private int cookingTime; //In minutes
 
-    public Dish(String name, DishType dishType, double price, double cookingTime) {
+    public Dish(String name, DishType dishType, double price, int cookingTime) {
         this.name = name;
         this.dishType = dishType;
         this.price = price;
@@ -46,7 +46,7 @@ public class Dish {
         return cookingTime;
     }
 
-    public void setCookingTime(double cookingTime) {
+    public void setCookingTime(int cookingTime) {
         this.cookingTime = cookingTime;
     }
